requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
django-ckeditor==5.7.1
django-downloadview==1.9
django-js-asset==1.2.2
django-paypal==1.0.0
idna==2.8
Pillow==6.1.0
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
